cognitive-care-scripts/requirements.txt
annotated-types==0.7.0
av==12.3.0
blis==0.7.11
catalogue==2.0.10
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.20.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.5
conllu==6.0.0
ctranslate2==4.5.0
cymem==2.0.8
en_core_sci_lg @ https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_lg-0.5.4.tar.gz#sha256=a2bfde10824958d51cd016b4142563373d553e24767d79b1abf8699596e31a5b
faster-whisper==1.0.3
filelock==3.13.1
flatbuffers==24.3.25
fsspec==2024.2.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
langcodes==3.4.1
language_data==1.2.0
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nmslib-metabrainz==2.1.3
numpy==1.26.4
onnxruntime==1.20.0
packaging==24.2
pillow==10.2.0
preshed==3.0.9
protobuf==5.28.3
psutil==6.1.0
psycopg2==2.9.10
pybind11==2.13.6
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyreadline3==3.5.4
pysbd==0.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
scispacy==0.5.5
shellingham==1.5.4
smart-open==7.0.5
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.13.1
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchvision==0.20.1+cu124
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.9.0
urllib3==2.2.3
wasabi==1.1.3
weasel==0.4.1
wrapt==1.16.0
